Struct prime::PrimeSet

pub struct PrimeSet {
    // some fields omitted
}

Prime number set

Methods

impl PrimeSet

fn new() -> PrimeSet

Create a new prime number generator.

fn new_empty() -> PrimeSet

Create a new prime number generator with empty buffers.

fn with_capacity(capacity: usize) -> PrimeSet

Create a new prime number generator with specifying buffer capacity.

fn nth(&self, n: usize) -> u64

Get nth prime.

Example

use prime::PrimeSet;
let ps = PrimeSet::new();
assert_eq!(2, ps.nth(0));
assert_eq!(3, ps.nth(1));
assert_eq!(5, ps.nth(2));
assert_eq!(743, ps.nth(131));

fn iter<'a>(&'a self) -> Nums

An iterator visiting all prime numbers in ascending order.

Example

use prime::PrimeSet;
let mut it = PrimeSet::new().iter();
assert_eq!(Some(2), it.next());
assert_eq!(Some(3), it.next());
assert_eq!(Some(5), it.next());
assert_eq!(Some(7), it.next());

fn contains(&self, n: u64) -> bool

Return true if the given number is prime.

fn combination(&self, n: u64, r: u64) -> u64

Calculates the combination of the number